Web Application to manage drivers/riders
This project is a part of my internship that I carried out in the MSME
Infinity Advertisement, under the aegis of PHD Chamber of Commerce,
University of Delhi.
Through this app employees of the company can generate requests for
their work requiring company drivers.
Requests are transferred to the manager who then allocates those tasks
to drivers.
Employees can track the progress of their requested tasks.
This app also maintains record of all the tasks, shows the amount of
work allocated to the drivers, and generate excel files, bar graphs etc
representing the statistics of the company to measure the requirement of
drivers in all their departments.
It also serves as a measure to analyse the performance of the drivers.